path: alterado path do arquivo para domain
