+++ components/layout/screenLayout/domain/types.ts
@@ -1,5 +1,5 @@
 import React, { ReactNode } from 'react';
-import { MenuItem } from '../../sidebar/types/types';
+import { MenuItem } from '../../sidebar/domain/types';
 
 export interface LayoutProps {
     children: ReactNode;
