+++ components/modal/customModal/domain/types.ts
