+++ components/modal/customModal/index.tsx
@@ -5,8 +5,8 @@ import { useTranslation } from 'react-i18next';
 import { Form, DatePicker, Card, Divider, Space, Row, Col } from "antd";
 import { InputField, Button, SelectComponent } from '../../base-components';
 import moment from 'moment';
-import { validFields } from '../../../helpers/utils';
-import { ModalProps, Field, FormData } from './types/types';
+import { validFields } from '../../../helpers/validFields';
+import { ModalProps, Field, FormData } from './domain/types';
 import { customStyles } from './constants/constants';
 
 export const CustomModal: React.FC<ModalProps> = ({ closeModal, onSave, fields, contentLabel, currentItem }) => {
