+++ infra/domain/type.ts
